refactor: share ActivityResultAwaiter across AndroidX activities
ReactiveAppCompatActivity and ReactiveFragmentActivity each nested an identical
ActivityResultAwaiter (flagged as duplicated by Sonar), differing only in the
Intent nullability of the result tuple. Lift it into a single shared
ActivityResultAwaiter and unify ReactiveFragmentActivity's ActivityResult /
StartActivityForResultAsync to Intent? (matching ReactiveAppCompatActivity) —
a correctness fix, since activity results can legitimately carry a null intent
(its OnActivityResult no longer throws on null data). PublicAPI baselines
updated for the net10/net11 android TFMs; both build clean.
