+++ src/ReactiveUI.AndroidX/ReactiveFragmentActivity.cs
@@ -23,7 +23,7 @@ public class ReactiveFragmentActivity : FragmentActivity, IReactiveObject,
     private readonly Signal<RxVoid> _deactivated = new();
 
     /// <summary>The subject that signals activity results.</summary>
-    private readonly Signal<(int requestCode, Result result, Intent intent)> _activityResult = new();
+    private readonly Signal<(int requestCode, Result result, Intent? intent)> _activityResult = new();
 
     /// <inheritdoc/>
     public event PropertyChangingEventHandler? PropertyChanging;
@@ -49,7 +49,7 @@ public class ReactiveFragmentActivity : FragmentActivity, IReactiveObject,
     public IObservable<RxVoid> Deactivated => _deactivated;
 
     /// <summary>Gets the activity result.</summary>
-    public IObservable<(int requestCode, Result result, Intent intent)> ActivityResult =>
+    public IObservable<(int requestCode, Result result, Intent? intent)> ActivityResult =>
         _activityResult;
 
     /// <inheritdoc/>
@@ -65,7 +65,7 @@ public class ReactiveFragmentActivity : FragmentActivity, IReactiveObject,
     /// <param name="intent">The intent.</param>
     /// <param name="requestCode">The request code.</param>
     /// <returns>A task with the result and intent.</returns>
-    public Task<(Result result, Intent intent)> StartActivityForResultAsync(Intent intent, int requestCode)
+    public Task<(Result result, Intent? intent)> StartActivityForResultAsync(Intent intent, int requestCode)
     {
         // NB: It's important that we set up the subscription *before* we
         // call ActivityForResult
@@ -79,7 +79,7 @@ public class ReactiveFragmentActivity : FragmentActivity, IReactiveObject,
     /// <param name="type">The type.</param>
     /// <param name="requestCode">The request code.</param>
     /// <returns>A task with the result and intent.</returns>
-    public Task<(Result result, Intent intent)> StartActivityForResultAsync(Type type, int requestCode)
+    public Task<(Result result, Intent? intent)> StartActivityForResultAsync(Type type, int requestCode)
     {
         // NB: It's important that we set up the subscription *before* we
         // call ActivityForResult
@@ -106,8 +106,6 @@ public class ReactiveFragmentActivity : FragmentActivity, IReactiveObject,
     /// <inheritdoc/>
     protected override void OnActivityResult(int requestCode, Result resultCode, Intent? data)
     {
-        ArgumentExceptionHelper.ThrowIfNull(data);
-
         base.OnActivityResult(requestCode, resultCode, data);
         _activityResult.OnNext((requestCode, resultCode, data));
     }
@@ -123,82 +121,5 @@ public class ReactiveFragmentActivity : FragmentActivity, IReactiveObject,
         }
 
         base.Dispose(disposing);
-    }
-
-    /// <summary>
-    /// Awaits the first <see cref="ActivityResult"/> that matches a request code, then completes the task and
-    /// detaches. A fused, allocation-light replacement for <c>Where(...).Select(...).FirstAsync().ToTask()</c>:
-    /// it is its own observer, settles exactly once, and unsubscribes on completion.
-    /// </summary>
-    private sealed class ActivityResultAwaiter
-        : IObserver<(int requestCode, Result result, Intent intent)>, IDisposable
-    {
-        /// <summary>The request code this awaiter is waiting for.</summary>
-        private readonly int _requestCode;
-
-        /// <summary>Completion source backing the returned task.</summary>
-        private readonly TaskCompletionSource<(Result result, Intent intent)> _completion = new();
-
-        /// <summary>Holds the source subscription so it can be torn down once settled.</summary>
-        private readonly OnceDisposable _subscription = new();
-
-        /// <summary>Guards against settling more than once.</summary>
-        private int _settled;
-
-        /// <summary>Initializes a new instance of the <see cref="ActivityResultAwaiter"/> class.</summary>
-        /// <param name="requestCode">The request code to await.</param>
-        private ActivityResultAwaiter(int requestCode) => _requestCode = requestCode;
-
-        /// <summary>Subscribes to the activity-result stream and returns a task for the first matching result.</summary>
-        /// <param name="source">The activity-result stream.</param>
-        /// <param name="requestCode">The request code to await.</param>
-        /// <returns>A task that completes with the result and intent of the first matching activity result.</returns>
-        public static Task<(Result result, Intent intent)> Await(
-            IObservable<(int requestCode, Result result, Intent intent)> source,
-            int requestCode)
-        {
-            var awaiter = new ActivityResultAwaiter(requestCode);
-            awaiter._subscription.Disposable = source.Subscribe(awaiter);
-            return awaiter._completion.Task;
-        }
-
-        /// <inheritdoc/>
-        public void OnNext((int requestCode, Result result, Intent intent) value)
-        {
-            if (value.requestCode != _requestCode || Interlocked.Exchange(ref _settled, 1) != 0)
-            {
-                return;
-            }
-
-            _ = _completion.TrySetResult((value.result, value.intent));
-            Dispose();
-        }
-
-        /// <inheritdoc/>
-        public void OnError(Exception error)
-        {
-            if (Interlocked.Exchange(ref _settled, 1) != 0)
-            {
-                return;
-            }
-
-            _ = _completion.TrySetException(error);
-            Dispose();
-        }
-
-        /// <inheritdoc/>
-        public void OnCompleted()
-        {
-            if (Interlocked.Exchange(ref _settled, 1) != 0)
-            {
-                return;
-            }
-
-            _ = _completion.TrySetCanceled();
-            Dispose();
-        }
-
-        /// <inheritdoc/>
-        public void Dispose() => _subscription.Dispose();
     }
 }
